Emit a warning instead of throwing an exception
Prior this commit, an UnsupportedOperationException was thrown when
a test module descriptor is present and a main module descriptor is
absent.
Now only a warning is emitted informing the user to either fix the
modular setup or to allow users ignoring the warning. They might
want to test that specific setup after all.