Improve "Older projects with module-info" chapter
The previous suggestion lead to all source code being compiled twice, once with Java 9 and once with Java 6.
This is unnecessarily slow. The new suggestion first compiles everything with Java 6 (except the module-info file)
and then compiles *only* the module-info file with Java 9. Since all the other files are already compiled at that point,
javac will be able to reference them from the classpath. This significantly speeds up compilation.

I've also put the "minimum Java 6" section at the top, as that is the more common case. Instructions for Java 5
are now at the bottom of the page. This makes it more likely for people to copy-paste only what they really need.