[MCOMPILER-398] Simplify logic around scanning for includes/excludes
The 'getSourceInclusionScanner' method can be simplified quite considerably.

In the case where there are no includes or excludes, we can simply reuse the
empty set of excludes we have rather than pulling an empty set out of
Collections. This then makes both constructor calls to
SimpleSourceInclusionScanner the same, so they can be merged.

For the case where there are no includes, there are two paths. When there are
also no excludes, we construct a singleton set of the default include pattern
and assign it to includes. In the other case, where we have excludes but no
includes, we add the defaultIncludePattern to the empty set of includes we
have. These are equivalent operations, so merge them as well.

i.e. have some default includes but allow this to be overridden and also allow
excludes to be specified.