[MCOMPILER-447] Unable to compile modularized test for a multi-version jar that is 	        not a java module by default
The following changes resolves two different issues.  The first is the case that
the we are creating a multi-version jar file that by default is not modularized.
If tests are created that will be testing the modularized part of the multi-version
jar file.  Then the build was generating an UnsupportedOperationException because
it assumed that we were using a non-modularized jar file with a modularized test.

The second issue was related to the fact that the test are using dependency jars
that are not modularized.  The generated javac command was only producing the
required module path and nulled out the classpath.  This was causing the building
of the test sources to fail.  The was caused because those non-modularized jar
files were not part of the classpath.

The current set of changes solves the first issue by determining if the project
jar file is a multi-version jar file.  If it is, it will then set the main module
descriptor to the latest version less than or equal to the defined release
attribute.

It solves the second issue by adding all non-modularized jar files to the
class path.  It does not remove them from the defined module path to be
consistent with how it was working.

Finally an integration test was created that reproduced the above issues.