[MCOMPILER-447] Unable to compile modularized test for a multi-version jar that is                         not a java module by default
As per Ben suggestion, refactored code into a method that is called.   The method
will check if a given path is a jar file.  If it is, it will then check to see if
it contains the module-info.class file.  If it doesn't then it will add the jar
file to the build class path.