[MCOMPILER-592] don't validate the project.build.outputTimestamp
The outputTimestamp is checked to be not shorter than one char or to be a digit.
The parsing of the outputTimestamp in the maven-jar-plugin is more complicated.
But it's not needed to be done in the compiler plugin at all so it would be safer to remove it.
We already had an IndexOutOfBoundsException when the string is empty because of the useless validation.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>