Rewrite the plugin using standard `javax.tools` API instead of Plexus. (#271)
* Rewrite the plugin using standard `javax.tools` API instead of Plexus.
Other aspects that are rewritten include incremental build mechanism
and how the overwriting of `module-info.java` in tests is handled.
For details and impact on users, see pull request description at
https://github.com/apache/maven-compiler-plugin/pull/271

* Temporarily disable the MCOMPILER-346 integration test.
It will be re-enabled after Maven 4.0.0-beta-6 release.

https://github.com/apache/maven/pull/1865

* Remove the replacement of `/` by `\` on the Windows platform in GLOB syntax.
It appears that the GLOB matcher expects `/` even on Windows.

* Add a log at the error level when the compiler threw an exception instead of returning the Boolean `false` value.
Adjust what is the cause and what is the suppressed exception.

* Set the `maven-plugin-testing-harness` version to 4.0.0-beta-2.

* Dummy change for forcing a new CI build.