Improve DeltaList behavior with large codebase
I have been experimenting with a large code base to try to improve
Quarkus build time and I stumbled upon the fact that DeltaList was very
inefficient in these cases.

In this commit, I keep the ordering even if not actually important. I
will push further optimizations in a second commit that we can discuss
further.

Fixes #334