Fix multi-release support when module-info is not in the base class (#948)
The plugin needs to be better at detecting whether to put the result targeting previous Java releases on the class-path or module-path.
This change requires better detection of `module-info.class` or `module-info.java` files which may be in non-straightforward directories.
More appropriate use of module-path also implies the use of `--module-source-path` instead of `--source-path` when patching a Java module,
even if Oracle's documentation of `javac` said that the use of `--module-source-path` option is not mandatory when there is only one module.
The plugin uses `--module-source-path` because `javac` seems to not always find `module-info` when it is located in a non-obvious directory.

`--module-source-path` causes `javac` to create an extra directory with the module name,
which is not the directory layout that Maven 3 users and other Maven plugins are used to.
The Maven Compiler Plugin was redirecting the `javac` output using a symbolic link,
but it appeared to be sometime a source of confusion because of files in two places.
This commit replaces the symbolic link by temporary renaming of directories.

Note: all this complexity is an effort to mimic the Maven 3 behaviour for compatibility reasons.
This complexity does not apply when the new `<sources>` element of Maven 4 is used,
because we accept that `javac` produces a slightly different dirctory layout in such case.