Build projects that are both multi-module and multi-release (#959)
Build projects that are both multi-module and multi-release.
Before this commit, a project could be either multi-module or multi-release, but not both in same time.
As a side effect of the debugging work, this commit also fixes the content of `target/javac.args` file
(before this commit, the source files for all versions were mixed in the same `target/javac.args` file).