When building a project which is both multi-module multi-release, rename the `META-INF/versions` subdirectory as `versions-modular`. Therefore, the full pattern of the path to the class files become `META-INF/versions-modular/${release}/${module}.
Rational: the "${module}" directory in above-cited path is inserted
by `javac` when module hierarchy is used. We have no control on that.
However, there is no multi-module JAR file format as of November 2025.
Therefore, dispatching the files for different modules in each JAR file
would be the work of the Maven JAR plugin. A prototype is already ready,
but experience with it shows that it is difficult to guess whether the
`META-INF/versions/`${release}` directory is expected to be followed by
a `${module}` subdirectory.

Renaming `versions` to anything else would avoid that ambiguity.
We propose "versions-modular" as a shortcut of "versions for a modular project".
The name is not very important because this directory will not appear in any JAR
or any other file to be distributed. Only its content is dispatched in JAR files.